Add fht as a submodule
Change the include to take into account the new path of fht
Add in the setup.py the instruction taken from the fht setup.py to compile it. Move the numpy import to be able to import it during compilation (before, if numpy was not present, it failed). Better for `pip install -e .`. Not sufficient for `python setup.py install`. See https://github.com/openspyrit/spyrit/blob/master/spyrit/learning/nets.py#L444 for more details.
Change in setup.py the instruction to compile fht for mac: add openmp flag.
Modify Readme to take into account submodule while cloning
Add fht symbolic link to the repository: as fht is a submodule, I do not want to modify it. So to import the functions from fht, it's easier with a symbolic link to the fht repository.
This symbolic link does not work for windows compilation (it works for execution). So I keep the full path to fht module in setup.py